+++ engine/language-server/src/main/scala/org/enso/languageserver/requesthandler/refactoring/RenameProjectHandler.scala
@@ -1,14 +1,18 @@
 package org.enso.languageserver.requesthandler.refactoring
 
-import java.util.UUID
-
 import akka.actor.{Actor, ActorRef, Cancellable, Props}
 import com.typesafe.scalalogging.LazyLogging
 import org.enso.jsonrpc._
-import org.enso.languageserver.refactoring.RefactoringApi.RenameProject
+import org.enso.languageserver.refactoring.RefactoringApi.{
+  ProjectRenameFailed,
+  RenameProject
+}
+import org.enso.languageserver.refactoring.RefactoringProtocol
 import org.enso.languageserver.requesthandler.RequestTimeout
 import org.enso.languageserver.util.UnhandledLogging
 import org.enso.polyglot.runtime.Runtime.Api
+
+import java.util.UUID
 
 import scala.concurrent.duration.FiniteDuration
 
@@ -52,8 +56,23 @@ class RenameProjectHandler(timeout: FiniteDuration, runtimeConnector: ActorRef)
       replyTo ! ResponseError(Some(id), Errors.RequestTimeout)
       context.stop(self)
 
-    case Api.Response(_, Api.ProjectRenamed(_, _)) =>
+    case Api.Response(
+          _,
+          Api.ProjectRenamed(oldNormalizedName, newNormalizedName, newName)
+        ) =>
+      context.system.eventStream.publish(
+        RefactoringProtocol.ProjectRenamedNotification(
+          oldNormalizedName,
+          newNormalizedName,
+          newName
+        )
+      )
       replyTo ! ResponseResult(RenameProject, id, Unused)
+      cancellable.cancel()
+      context.stop(self)
+
+    case Api.Response(_, Api.ProjectRenameFailed(oldName, newName)) =>
+      replyTo ! ResponseError(Some(id), ProjectRenameFailed(oldName, newName))
       cancellable.cancel()
       context.stop(self)
   }